+++ ErsatzTV.Application/Streaming/PlayoutItemProcessModel.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using System.Diagnostics;
+
+namespace ErsatzTV.Application.Streaming
+{
+    public record PlayoutItemProcessModel(Process Process, DateTimeOffset Until);
+}