+++ ErsatzTV.Application/Streaming/Queries/GetPlayoutItemProcessByChannelNumber.cs
@@ -1,10 +1,17 @@
-﻿namespace ErsatzTV.Application.Streaming.Queries
+﻿using System;
+
+namespace ErsatzTV.Application.Streaming.Queries
 {
     public record GetPlayoutItemProcessByChannelNumber : FFmpegProcessRequest
     {
-        public GetPlayoutItemProcessByChannelNumber(string channelNumber, string mode, bool startAtZero) : base(
+        public GetPlayoutItemProcessByChannelNumber(
+            string channelNumber,
+            string mode,
+            DateTimeOffset now,
+            bool startAtZero) : base(
             channelNumber,
             mode,
+            now,
             startAtZero)
         {
         }