+++ ErsatzTV.Core/Interfaces/FFmpeg/IHlsSessionWorker.cs
@@ -1,0 +1,7 @@
+﻿namespace ErsatzTV.Core.Interfaces.FFmpeg
+{
+    public interface IHlsSessionWorker
+    {
+        void Touch();
+    }
+}