+++ ErsatzTV/Services/FFmpegWorkerService.cs
@@ -1,16 +1,11 @@
 ﻿using System;
-using System.Diagnostics;
 using System.IO;
 using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 using ErsatzTV.Application;
 using ErsatzTV.Application.Streaming.Commands;
-using ErsatzTV.Application.Streaming.Queries;
-using ErsatzTV.Core;
 using ErsatzTV.Core.Interfaces.FFmpeg;
-using LanguageExt;
-using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -21,20 +16,17 @@ namespace ErsatzTV.Services
     public class FFmpegWorkerService : BackgroundService
     {
         private readonly ChannelReader<IFFmpegWorkerRequest> _channel;
-        private readonly ChannelWriter<IFFmpegWorkerRequest> _channelWriter;
         private readonly ILogger<FFmpegWorkerService> _logger;
         private readonly IFFmpegSegmenterService _ffmpegSegmenterService;
         private readonly IServiceScopeFactory _serviceScopeFactory;
 
         public FFmpegWorkerService(
             ChannelReader<IFFmpegWorkerRequest> channel,
-            ChannelWriter<IFFmpegWorkerRequest> channelWriter,
             IServiceScopeFactory serviceScopeFactory,
             ILogger<FFmpegWorkerService> logger,
             IFFmpegSegmenterService ffmpegSegmenterService)
         {
             _channel = channel;
-            _channelWriter = channelWriter;
             _serviceScopeFactory = serviceScopeFactory;
             _logger = logger;
             _ffmpegSegmenterService = ffmpegSegmenterService;
@@ -49,7 +41,7 @@ namespace ErsatzTV.Services
                 try
                 {
                     using IServiceScope scope = _serviceScopeFactory.CreateScope();
-                    IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
+                    // IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
 
                     switch (request)
                     {
@@ -57,52 +49,6 @@ namespace ErsatzTV.Services
                             foreach (DirectoryInfo parent in Optional(Directory.GetParent(touchFFmpegSession.Path)))
                             {
                                 _ffmpegSegmenterService.TouchChannel(parent.Name);
-                            }
-                            break;
-                        case CleanUpFFmpegSessions:
-                            _ffmpegSegmenterService.CleanUpSessions();
-                            break;
-                        case StartFFmpegSession startFFmpegSession:
-                            _logger.LogInformation(
-                                "Starting ffmpeg session for channel {Channel}",
-                                startFFmpegSession.ChannelNumber);
-
-                            if (!_ffmpegSegmenterService.ProcessExistsForChannel(startFFmpegSession.ChannelNumber))
-                            {
-                                var req = new GetPlayoutItemProcessByChannelNumber(
-                                    startFFmpegSession.ChannelNumber,
-                                    "segmenter",
-                                    startFFmpegSession.StartAtZero);
-                                Either<BaseError, Process> maybeProcess = await mediator.Send(req, cancellationToken);
-                                maybeProcess.Match(
-                                    process =>
-                                    {
-                                        if (_ffmpegSegmenterService.TryAdd(startFFmpegSession.ChannelNumber, process))
-                                        {
-                                            _logger.LogDebug(
-                                                "ffmpeg hls arguments {FFmpegArguments}",
-                                                string.Join(" ", process.StartInfo.ArgumentList));
-
-                                            process.Start();
-                                            process.EnableRaisingEvents = true;
-                                            process.Exited += (_, _) =>
-                                            {
-                                                if (process.ExitCode == 0)
-                                                {
-                                                    _channelWriter.TryWrite(
-                                                        new StartFFmpegSession(startFFmpegSession.ChannelNumber, true));
-                                                }
-                                                else
-                                                {
-                                                    _logger.LogDebug(
-                                                        "hls segmenter for channel {Channel} exited with code {ExitCode}",
-                                                        startFFmpegSession.ChannelNumber,
-                                                        process.ExitCode);
-                                                }
-                                            };
-                                        }
-                                    },
-                                    _ => { });
                             }
 
                             break;
@@ -113,9 +59,6 @@ namespace ErsatzTV.Services
                      _logger.LogWarning(ex, "Failed to handle ffmpeg worker request");
                 }
             }
-            
-            // kill any running processes after cancellation
-            _ffmpegSegmenterService.KillAll();
         }
     }
 }