+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Changed
-- Reduce initial `HLS Segmenter` delay from 5 seconds to 3 seconds
+- Remove forced initial delay from `HLS Segmenter` streaming mode
 
 ## [0.1.0-alpha] - 2021-10-08
 ### Added